worker-python/requirements.txt
pandas
psycopg2-binary
httpx
tqdm
ijson
google-cloud-billing
